+++ tests/dataset/test_serialize_json.py
@@ -1,5 +1,5 @@
-from tests.data.test_common import _assert_tar_file_contents
-from unifair.data.json import JsonDataset, JsonDatasetToTarFileSerializer
+from tests.dataset.test_common import _assert_tar_file_contents
+from unifair.dataset.json import JsonDataset, JsonDatasetToTarFileSerializer
 
 
 def test_json_dataset_to_tar_file_serializer_single_obj_type():
