+++ unifair/dataset/pandas.py
@@ -4,10 +4,10 @@ from typing import Dict, IO, Union
 import pandas as pd
 from pydantic import validator
 
-from unifair.data import (create_dataset_from_tarfile,
-                          create_tarfile_from_dataset,
-                          Dataset,
-                          validate)
+from unifair.dataset import (create_dataset_from_tarfile,
+                             create_tarfile_from_dataset,
+                             Dataset,
+                             validate)
 
 
 class PandasDataset(Dataset):
